\name{Bfiat-package}
\alias{Bfiat-package}
\alias{Bfiat}
\docType{package}
\title{
  Bottom Fishing Impact Assessment Tools
}
\description{
  Functions to estimate the impact of bottom disturbances on sediment ecosystems.
    \if{html}{\figure{Logo_fishing.jpg}{options: width=80 alt="Logo" align='right'}}
\if{latex}{\figure{Logo_fishing.jpg}{options: width=0.4in}}
}

\author{
  Karline Soetaert, Olivier Beauchard
}
\references{
Beauchard, O., Murray S.A. Thompson, Kari Elsa Ellingsen, Gerjan Piet, Pascal Laffargue, Karline Soetaert, 2023. Assessing sea floor functional diversity and vulnerability. 
Marine Ecology Progress Serie v708, p21-43, https://www.int-res.com/abstracts/meps/v708/p21-43/

Data product created under the European Marine Observation Data Network (EMODnet) Biology Phase V, under the NECCTON project (New Copernicus Capability for Trophic Ocean Networks), and under the OceanICU (understanding Ocean Carbon) project.
}

\keyword{ package }
